Build and publish the docket search index in the new pipeline
The new pipeline did not rebuild docket_search.json.gz, so cutting the
daily cron over to it would have let the in-browser search index go
stale. Wire it into RegulationsPipeline alongside the other
post-processing:

- Move the search-index builder to transforms/build_search_index.py (the
  canonical home for dataset transforms); pipeline/build_search_index.py
  becomes a back-compat shim that preserves the `python -m` entry point.
- RegulationsPipeline rebuilds the index during post-processing when
  dockets changed, and uploads it to R2 when upload is on — matching the
  legacy pipeline's gating (last batch only, dockets-changed only).
- Add r2.upload_file() for single-file publishes.
